requirements.txt

aiogram>=3.20.0
google-api-python-client>=2.170.0
google-auth>=2.40.2
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.2
python-dotenv>=1.1.0
requests>=2.32.3
Flask>=3.0.0
aiogram
